+++ :5001/0/fbfb5b4ef32c4d2c9853cffab3d0f044/meta.yaml
@@ -1,0 +1,15 @@
+artifact_uri: :5001/0/fbfb5b4ef32c4d2c9853cffab3d0f044/artifacts
+end_time: null
+entry_point_name: ''
+experiment_id: '0'
+lifecycle_stage: active
+name: ''
+run_id: fbfb5b4ef32c4d2c9853cffab3d0f044
+run_uuid: fbfb5b4ef32c4d2c9853cffab3d0f044
+source_name: ''
+source_type: 4
+source_version: ''
+start_time: 1584822129318
+status: 1
+tags: []
+user_id: amritbaveja
